requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
aiofiles==24.1.0
openpyxl==3.1.5
reportlab==4.2.2
pandas==2.2.3
alembic==1.13.3
python-dotenv==1.0.1
# HTTP client (Paystack API calls)
httpx==0.27.2
# PostgreSQL driver (used on Heroku; SQLite used locally)
psycopg2-binary==2.9.9
# Queue / workers
celery[redis]==5.4.0
redis==5.1.1
flower==2.0.1
